integration tests

+++ app/tests/test_dynamo.py
@@ -41,7 +41,8 @@ class TestWriteResult:
         item_written = mock_table.put_item.call_args.kwargs["Item"]
         assert item_written["sql"] == "SELECT COUNT(*) FROM t"
         assert item_written["rows"] == '[{"n": 1}]'
-        assert isinstance(item_written["latency_ms"], float)
+        from decimal import Decimal
+        assert isinstance(item_written["latency_ms"], Decimal)
 
     def test_retries_on_failure_then_raises(self) -> None:
         mock_table = MagicMock()
